// backend/src/modules/tasks/dto/task-enums.dto.ts
export enum TaskStatusDto {
  BACKLOG = 'backlog',
  IN_PROGRESS = 'inProgress',
  DONE = 'done',
  MISSED = 'missed',
  DELETED = 'deleted',
}

export enum TaskDifficultyDto {
  EASY = 'easy',
  MEDIUM = 'medium',
  HARD = 'hard',
}

export enum TaskPriorityDto {
  LOW = 'low',
  MEDIUM = 'medium',
  HIGH = 'high',
}